Rewrite service worker registrar in TS
